TSIS/TSIS3/requirements.txt
pygame>=2.5.0
psycopg2-binary>=2.9.9
